Make JsonReader tests reusable in the com.google.gson.stream package
Make the JsonReader tests reusable so that we can run the same set of tests
with differently configured JsonReaders.
Note that this change does not expose a new API, which could be subclassed
by 3rd party classes, because the AbstractJsonReaderTest has package-private
visibility and the JsonReaderTest is final.
